selectionner journee ajax

+++ src/Controller/Admin/UserController.php
@@ -3,14 +3,16 @@
 namespace App\Controller\Admin;
 
 use App\Entity\User;
-use App\Form\SearchUserType;
 use App\Form\UserType;
+use App\Form\SearchUserType;
 use App\Form\UserUpdateFormType;
 use App\Repository\UserRepository;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Security\Csrf\CsrfToken;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 
@@ -163,8 +165,14 @@ class UserController extends AbstractController
     }
 
     #[Route("/delete/{id}", name: "delete")]
-    public function delete(User $user, ManagerRegistry $doctrine): Response
+    public function delete(User $user, ManagerRegistry $doctrine, Request $request, CsrfTokenManagerInterface $csrfTokenManager): Response
     {
+        $token = new CsrfToken('delete', $request->query->get('_csrf_token'));
+
+        if (!$csrfTokenManager->isTokenValid($token)) {
+            throw $this->createAccessDeniedException('Token CSRF invalide');
+        }
+
         $em = $doctrine->getManager();
         $em->remove($user);
         $em->flush();
